audit(phase-1): harden MySQL parity and money boundaries

+++ backend/Data/MarketingSchema.cs
@@ -75,8 +75,8 @@ public static class MarketingSchema
             audience_summary TEXT, geography TEXT, language VARCHAR(10),
             landing_page_id INTEGER, promotion_id INTEGER,
             start_date TEXT, end_date TEXT,
-            total_budget REAL DEFAULT 0, budget_currency VARCHAR(8) DEFAULT 'USD',
-            alloc_linkedin REAL DEFAULT 0, alloc_google REAL DEFAULT 0, alloc_meta REAL DEFAULT 0,
+            total_budget DECIMAL(12,2) DEFAULT 0, budget_currency VARCHAR(8) DEFAULT 'USD',
+            alloc_linkedin DECIMAL(12,2) DEFAULT 0, alloc_google DECIMAL(12,2) DEFAULT 0, alloc_meta DECIMAL(12,2) DEFAULT 0,
             conversion_goal TEXT, status VARCHAR(30) DEFAULT 'draft', approval_status VARCHAR(30) DEFAULT 'draft',
             created_by INTEGER, created_at TEXT DEFAULT (datetime('now')), updated_at TEXT DEFAULT (datetime('now')))");
         db.Exec("CREATE INDEX IF NOT EXISTS ix_mkt_camp_status ON mkt_campaigns(status)");
@@ -86,7 +86,7 @@ public static class MarketingSchema
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             campaign_id INTEGER NOT NULL, platform_code VARCHAR(40) NOT NULL, connection_id INTEGER,
             provider_campaign_id TEXT, name TEXT, objective TEXT, campaign_type TEXT,
-            daily_budget REAL, lifetime_budget REAL, bid_strategy TEXT,
+            daily_budget DECIMAL(12,2), lifetime_budget DECIMAL(12,2), bid_strategy TEXT,
             targeting_json TEXT, landing_page_id INTEGER, lead_form_id INTEGER, conversion_id INTEGER,
             status VARCHAR(30) DEFAULT 'draft', provider_status TEXT, approval_status VARCHAR(30) DEFAULT 'draft',
             last_synced_at TEXT, created_at TEXT DEFAULT (datetime('now')), updated_at TEXT DEFAULT (datetime('now')))");
@@ -155,7 +155,7 @@ public static class MarketingSchema
             campaign_id INTEGER, connection_id INTEGER, name TEXT NOT NULL, objective TEXT,
             sender_identity TEXT, intro_message TEXT, steps_json TEXT, buttons_json TEXT,
             audience_id INTEGER, landing_page_id INTEGER, lead_form_id INTEGER,
-            daily_budget REAL, lifetime_budget REAL, start_date TEXT, end_date TEXT, frequency_cap TEXT,
+            daily_budget DECIMAL(12,2), lifetime_budget DECIMAL(12,2), start_date TEXT, end_date TEXT, frequency_cap TEXT,
             status VARCHAR(30) DEFAULT 'draft', approval_status VARCHAR(30) DEFAULT 'draft',
             provider_campaign_id TEXT, metrics_json TEXT,
             created_by INTEGER, created_at TEXT DEFAULT (datetime('now')), updated_at TEXT DEFAULT (datetime('now')))");
@@ -219,7 +219,7 @@ public static class MarketingSchema
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             platform_campaign_id INTEGER, list_name TEXT, keyword TEXT NOT NULL, match_type VARCHAR(16),
             kind VARCHAR(12) DEFAULT 'keyword',  -- keyword | negative
-            status VARCHAR(16) DEFAULT 'active', max_cpc REAL, created_by INTEGER, created_at TEXT DEFAULT (datetime('now')))");
+            status VARCHAR(16) DEFAULT 'active', max_cpc DECIMAL(18,6), created_by INTEGER, created_at TEXT DEFAULT (datetime('now')))");
 
         // ── Budget approvals ──
         db.Exec(@"CREATE TABLE IF NOT EXISTS mkt_budget_approvals(
@@ -264,8 +264,8 @@ public static class MarketingSchema
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             platform_campaign_id INTEGER, platform_code VARCHAR(40), day TEXT,
             impressions INTEGER DEFAULT 0, reach INTEGER DEFAULT 0, clicks INTEGER DEFAULT 0,
-            spend REAL DEFAULT 0, currency VARCHAR(8) DEFAULT 'USD',
-            leads INTEGER DEFAULT 0, conversions INTEGER DEFAULT 0, conversion_value REAL DEFAULT 0,
+            spend DECIMAL(18,6) DEFAULT 0, currency VARCHAR(8) DEFAULT 'USD',
+            leads INTEGER DEFAULT 0, conversions INTEGER DEFAULT 0, conversion_value DECIMAL(18,6) DEFAULT 0,
             video_views INTEGER DEFAULT 0, dedup_key VARCHAR(160),
             created_at TEXT DEFAULT (datetime('now')))");
         db.Exec("CREATE UNIQUE INDEX IF NOT EXISTS ux_mkt_metrics_dedup ON mkt_campaign_metrics(dedup_key)");
